+++ interfaces/ShardingStrategyInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace Flex\Banana\Interfaces;
+
+interface ShardingStrategyInterface {
+    public function addServer(string $group, string $server): void;
+    public function removeServer(string $group, string $server): void;
+    public function getServer(string $group, string $key): ?string;
+}